Improve the documentation & Optimize the navigation page & Add Submit-Button & Add setting-service & Add the CONSOLE enable for the map page
