+++ tools/InternetMap/backend/src/api/v1/main.ts
@@ -47,6 +47,16 @@ controller.getLoggers().forEach(logger => logger.setSettings({
     minLevel: 'warn'
 }));
 
+router.get('/env.js', (req, res, next) => {
+  const envVarsForFrontend = {
+    CONSOLE: process.env.CONSOLE,
+  };
+  res.setHeader('Content-Type', 'application/javascript');
+  res.send(`window.__ENV__ = ${JSON.stringify(envVarsForFrontend)}`);
+
+  next();
+});
+
 router.get('/network', async function (req, res, next) {
     var networks = await docker.listNetworks();
 
@@ -280,6 +290,9 @@ router.post('/container/vis/set', express.json(), async function (req, res, next
 
 router.ws('/console/:id', async function (ws, req, next) {
     try {
+        if (process.env.CONSOLE === 'false') {
+            throw Error('CONSOLE is not enabled');
+        }
         await socketHandler.handleSession(ws, req.params.id);
     } catch (e) {
         if (ws.readyState == 1) {
