+++ tools/InternetMap/frontend/public/dashboard.html
@@ -1,0 +1,101 @@
+<!DOCTYPE html>
+<html lang="en">
+
+<head>
+    <title>SEEDEMU Dashboard</title>
+</head>
+
+<body>
+    <nav class="navbar navbar-light fixed-top bg-light d-flex justify-content-between align-items-center">
+        <a class="navbar-brand" href="#">SEEDEMU Dashboard</a>
+        <ul class="navbar-nav mr-auto">
+            <li class="nav-item">
+                <a class="nav-link" href="#" data-toggle="modal" data-target="#help-modal">Help</a>
+            </li>
+        </ul>
+        <div class="d-flex">
+            <a class="p-2 text-dark" href="/map.html">Map</a>
+            <a class="p-2 text-dark" href="/plugin.html">Plugin</a>
+            <a class="p-2 text-dark" href="/index.html">Home</a>
+        </div>
+    </nav>
+    <div class="container-fluid">
+        <div>
+            <h2>Nodes</h2>
+            <table class="table table-sm" id="container-list">
+                <thead>
+                    <tr>
+                        <th></th>
+                        <th scope="col">ASN</th>
+                        <th scope="col">Name</th>
+                        <th scope="col">Type</th>
+                        <th scope="col">IP Address(es)</th>
+                        <th scope="col">Actions</th>
+                    </tr>
+                </thead>
+                <tbody>
+                </tbody>
+            </table>
+        </div>
+        <div class="mt-3">
+            <h2>Networks</h2>
+            <table class="table table-sm" id="network-list">
+                <thead>
+                    <tr>
+                        <th></th>
+                        <th scope="col">ASN (scope)</th>
+                        <th scope="col">Name</th>
+                        <th scope="col">Type</th>
+                        <th scope="col">Network Prefix</th>
+                        <th scope="col">Actions</th>
+                    </tr>
+                </thead>
+                <tbody>
+                </tbody>
+            </table>
+        </div>
+    </div>
+    <div class="console-area" id="console-area"></div>
+    <div class="taskbar hide" id="taskbar"></div>
+    <div class="modal fade" id="help-modal" tabindex="-1" role="dialog" aria-labelledby="help-modal" aria-hidden="true">
+        <div class="modal-dialog" role="document">
+          <div class="modal-content">
+            <div class="modal-header">
+              <h5 class="modal-title" id="help-modal-label">Help</h5>
+              <button type="button" class="close" data-dismiss="modal" aria-label="Close">
+                <span aria-hidden="true">&times;</span>
+              </button>
+            </div>
+            <div class="modal-body">
+                <p>
+                    <kbd><kbd>ctrl</kbd> + <kbd>alt</kbd> + <kbd>w</kbd></kbd>: close current window.
+                </p>
+                <p>
+                    <kbd><kbd>ctrl</kbd> + <kbd>alt</kbd> + <kbd>s</kbd></kbd>: minimize current window.
+                </p>
+                <p>
+                    <kbd><kbd>ctrl</kbd> + <kbd>alt</kbd> + <kbd>e</kbd></kbd>: toggle input-broadcast of current window.
+                </p>
+                <p>
+                    <kbd><kbd>ctrl</kbd> + <kbd>alt</kbd> + <kbd>f</kbd></kbd>: pop-out current window.
+                </p>
+                <p>
+                    <kbd><kbd>ctrl</kbd> + <kbd>alt</kbd> + <kbd>r</kbd></kbd>: reload current window.
+                </p>
+                <p>
+                    <kbd><kbd>ctrl</kbd> + <kbd>alt</kbd> + <kbd>d</kbd></kbd>: minimize all windows.
+                </p>
+                <p>
+                    <kbd><kbd>ctrl</kbd> + <kbd>alt</kbd> + <kbd>a</kbd></kbd>: toggle input-broadcast of all windows.
+                </p>
+            </div>
+            <div class="modal-footer">
+              <button type="button" class="btn btn-secondary" data-dismiss="modal">Close</button>
+            </div>
+          </div>
+        </div>
+      </div>
+    <script src='./dashboard.js'></script>
+</body>
+
+</html>
