+++ tools/InternetMap/frontend/public/index.html
@@ -2,99 +2,68 @@
 <html lang="en">
 
 <head>
-    <title>SEEDEMU Dashboard</title>
+    <title>SEEDEMU HOME</title>
 </head>
 
 <body>
-    <nav class="navbar navbar-light fixed-top bg-light d-flex justify-content-between align-items-center">
-        <a class="navbar-brand" href="#">SEEDEMU Dashboard</a>
-        <ul class="navbar-nav mr-auto">
-            <li class="nav-item">
-                <a class="nav-link" href="#" data-toggle="modal" data-target="#help-modal">Help</a>
-            </li>
-        </ul>
-        <div class="d-flex">
-            <a class="p-2 text-dark" href="/map.html">Map</a>
-            <a class="p-2 text-dark" href="/install.html">Install</a>
+<nav class="navbar navbar-light fixed-top bg-light d-flex justify-content-between align-items-center">
+    <a class="navbar-brand" href="#">INTERNET MAP</a>
+    <ul class="navbar-nav mr-auto">
+        <li class="nav-item">
+            <a class="nav-link" href="#" data-toggle="modal" data-target="#help-modal">Help</a>
+        </li>
+    </ul>
+    <div class="d-flex">
+        <a class="p-2 text-dark" href="/map.html">Map</a>
+        <a class="p-2 text-dark" href="/plugin.html">Plugin</a>
+        <a class="p-2 text-dark" href="/dashboard.html">Dashboard</a>
+    </div>
+</nav>
+<main role="main">
+    <section class="jumbotron text-center">
+        <div class="container">
+            <h1>INTERNET MAP</h1>
+            <p class="lead text-muted">This is a work-in-progress prototype of the internet-map..</p>
         </div>
-    </nav>
-    <div class="container-fluid">
-        <div>
-            <h2>Nodes</h2>
-            <table class="table table-sm" id="container-list">
-                <thead>
-                    <tr>
-                        <th></th>
-                        <th scope="col">ASN</th>
-                        <th scope="col">Name</th>
-                        <th scope="col">Type</th>
-                        <th scope="col">IP Address(es)</th>
-                        <th scope="col">Actions</th>
-                    </tr>
-                </thead>
-                <tbody>
-                </tbody>
-            </table>
-        </div>
-        <div class="mt-3">
-            <h2>Networks</h2>
-            <table class="table table-sm" id="network-list">
-                <thead>
-                    <tr>
-                        <th></th>
-                        <th scope="col">ASN (scope)</th>
-                        <th scope="col">Name</th>
-                        <th scope="col">Type</th>
-                        <th scope="col">Network Prefix</th>
-                        <th scope="col">Actions</th>
-                    </tr>
-                </thead>
-                <tbody>
-                </tbody>
-            </table>
+    </section>
+    <div class="album py-5 bg-light">
+        <div class="container">
+            <div class="row">
+                <div class="col-md-4">
+                    <div class="card mb-4 shadow-sm">
+                        <a class="home-item" href="map.html" data-toggle="tooltip" data-placement="top" title="The network topology diagram displays interconnection relationships between nodes and networks, along with auxiliary functions including filtering, search, settings, replay, and logging.">
+                            <img class="bd-placeholder-img card-img-top" width="100%" height="225" alt="" src="assets/map.png">
+                            <div class="card-body">
+                                <p class="card-text">The network topology diagram displays interconnection relationships between nodes and networks, along with auxiliary functions including filtering, search, settings, replay, and logging. </p>
+                            </div>
+                        </a>
+                    </div>
+                </div>
+                <div class="col-md-4">
+                    <div class="card mb-4 shadow-sm">
+                        <a class="home-item" href="dashboard.html" data-toggle="tooltip" data-placement="top" title="seedemu dashboard, displays current emulator nodes and networks.">
+                            <img class="bd-placeholder-img card-img-top" width="100%" height="225" alt="" src="assets/dashboard.png">
+                            <div class="card-body">
+                                <p class="card-text">seedemu dashboard, displays current emulator nodes and networks. </p>
+                            </div>
+                        </a>
+                    </div>
+                </div>
+                <div class="col-md-4">
+                    <div class="card mb-4 shadow-sm">
+                        <a class="home-item" href="plugin.html" data-toggle="tooltip" data-placement="top" title="Plugin installation page.">
+                            <img class="bd-placeholder-img card-img-top" width="100%" height="225" alt="" src="assets/plugin.png">
+                            <div class="card-body">
+                                <p class="card-text">Plugin installation page. </p>
+                            </div>
+                        </a>
+                    </div>
+                </div>
+            </div>
         </div>
     </div>
-    <div class="console-area" id="console-area"></div>
-    <div class="taskbar hide" id="taskbar"></div>
-    <div class="modal fade" id="help-modal" tabindex="-1" role="dialog" aria-labelledby="help-modal" aria-hidden="true">
-        <div class="modal-dialog" role="document">
-          <div class="modal-content">
-            <div class="modal-header">
-              <h5 class="modal-title" id="help-modal-label">Help</h5>
-              <button type="button" class="close" data-dismiss="modal" aria-label="Close">
-                <span aria-hidden="true">&times;</span>
-              </button>
-            </div>
-            <div class="modal-body">
-                <p>
-                    <kbd><kbd>ctrl</kbd> + <kbd>alt</kbd> + <kbd>w</kbd></kbd>: close current window.
-                </p>
-                <p>
-                    <kbd><kbd>ctrl</kbd> + <kbd>alt</kbd> + <kbd>s</kbd></kbd>: minimize current window.
-                </p>
-                <p>
-                    <kbd><kbd>ctrl</kbd> + <kbd>alt</kbd> + <kbd>e</kbd></kbd>: toggle input-broadcast of current window.
-                </p>
-                <p>
-                    <kbd><kbd>ctrl</kbd> + <kbd>alt</kbd> + <kbd>f</kbd></kbd>: pop-out current window.
-                </p>
-                <p>
-                    <kbd><kbd>ctrl</kbd> + <kbd>alt</kbd> + <kbd>r</kbd></kbd>: reload current window.
-                </p>
-                <p>
-                    <kbd><kbd>ctrl</kbd> + <kbd>alt</kbd> + <kbd>d</kbd></kbd>: minimize all windows.
-                </p>
-                <p>
-                    <kbd><kbd>ctrl</kbd> + <kbd>alt</kbd> + <kbd>a</kbd></kbd>: toggle input-broadcast of all windows.
-                </p>
-            </div>
-            <div class="modal-footer">
-              <button type="button" class="btn btn-secondary" data-dismiss="modal">Close</button>
-            </div>
-          </div>
-        </div>
-      </div>
-    <script src='./index.js'></script>
+</main>
+<script src='./index.js'></script>
 </body>
 
 </html>
