+++ tools/InternetMap/frontend/public/plugin.html
@@ -2,12 +2,12 @@
 <html lang="en">
 
 <head>
-    <title>SEEDEMU INSTALL</title>
+    <title>SEEDEMU PLUGIN</title>
 </head>
 
 <body>
     <nav class="navbar navbar-light fixed-top bg-light d-flex justify-content-between align-items-center">
-        <a class="navbar-brand" href="#">SEEDEMU Install</a>
+        <a class="navbar-brand" href="#">SEEDEMU PLUGIN</a>
         <ul class="navbar-nav mr-auto">
             <li class="nav-item">
                 <a class="nav-link" href="#" data-toggle="modal" data-target="#help-modal">Help</a>
@@ -15,12 +15,13 @@
         </ul>
         <div class="d-flex">
             <a class="p-2 text-dark" href="/map.html">Map</a>
-            <a class="p-2 text-dark" href="/index.html">Index</a>
+            <a class="p-2 text-dark" href="/dashboard.html">Dashboard</a>
+            <a class="p-2 text-dark" href="/index.html">Home</a>
         </div>
     </nav>
     <div class="container-fluid">
         <div>
-            <h2>Install</h2>
+            <h2>Plugin</h2>
             <table class="table table-sm" id="install-list">
                 <thead>
                     <tr>
@@ -108,7 +109,7 @@
           </div>
         </div>
       </div>
-    <script src='./install.js'></script>
+    <script src='./plugin.js'></script>
 </body>
 
 </html>
