+++ tools/InternetMap/frontend/src/dashboard/dashboard.ts
@@ -9,3 +9,15 @@ import 'datatables.net';
 import 'datatables.net-bs4';
 import 'datatables.net-select';
 import 'datatables.net-select-bs4';
+
+import { DashboardUi } from './ui';
+
+var ui = new DashboardUi({
+    containerListElementId: 'container-list', 
+    networkListElementId: 'network-list',
+    desktopElementId: 'console-area',
+    taskbarElementId: 'taskbar'
+});
+
+ui.loadContainers('/api/v1/container');
+ui.loadNetworks('/api/v1/network')
